chore(monorepo): add storybook and update stories
