+++ .env
@@ -1,1 +1,1 @@
-QUAD_WORD=""
+VITE_QUAD_WORD=""
