+++ .storybook/preview-head.html
@@ -1,8 +1,3 @@
-<link
-  rel="stylesheet"
-  href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css"
-/>
-<link
-  rel="stylesheet"
-  href="https://cdn.jsdelivr.net/npm/tailwindcss/dist/tailwind.min.css"
-/>
+<script>
+  window.global = window;
+</script>
