+++ package.json
@@ -18,21 +18,18 @@
   "scripts": {
     "build": "",
     "build-storybook": "build-storybook",
-    "format": "prettier ./ --write",
+    "format": "prettier . --write",
     "graduate": "",
-    "lint": "eslint ./ --ext .js --ignore-path ./.eslintignore",
-    "lint:fix": "eslint ./ --ext .js --quiet --fix --ignore-path ./.eslintignore",
+    "lint": "eslint .",
+    "lint:fix": "eslint . --quiet --fix",
     "pre-release": "",
     "prepare": "",
     "release": "",
     "start": "",
     "storybook": "start-storybook -p 6006",
-    "test": "npm run test --workspaces --if-present"
+    "test": "vitest"
   },
   "prettier": {
-    "plugins": [
-      "prettier-plugin-packagejson"
-    ],
     "semi": true,
     "singleQuote": true
   },
@@ -74,7 +71,17 @@
   },
   "devDependencies": {
     "@arcgis/core": "^4.18.1",
+    "@babel/core": "^7.18.5",
+    "@storybook/addon-actions": "^6.5.9",
+    "@storybook/addon-essentials": "^6.5.9",
+    "@storybook/addon-interactions": "^6.5.9",
+    "@storybook/addon-links": "^6.5.9",
+    "@storybook/builder-vite": "^0.1.36",
+    "@storybook/react": "^6.5.9",
+    "@storybook/testing-library": "^0.0.13",
+    "@types/arcgis-js-api": "^4.24.0",
     "@vitejs/plugin-react": "^1.3.2",
+    "babel-loader": "^8.2.5",
     "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-import": "^2.26.0",
@@ -87,5 +94,9 @@
     "sass": "^1.32.4",
     "vite": "^2.9.12",
     "vitest": "^0.15.2"
+  },
+  "dependencies": {
+    "react": ">=16.8.0",
+    "react-dom": ">=16.8.0"
   }
 }
