+++ packages/dart-board/package.json
@@ -35,9 +35,10 @@
   ],
   "scripts": {
     "build": "vite build --config ../../vite.config.js",
-    "test": "npm test"
+    "test": "vitest --config ../../vite.config.js"
   },
   "dependencies": {
+    "@arcgis/core": "^4.0.0",
     "@ugrc/utilities": "^1.3.1",
     "clsx": "^1.1.1",
     "reactstrap": "^8.8.1"
