+++ packages/layer-selector/src/LayerSelectorWithMap.stories.jsx
@@ -1,21 +1,27 @@
-import React, { useRef, useEffect, useState } from 'react';
-import LayerSelector from './LayerSelector';
-import { ModulesHelper } from '../../../test-helpers/storyHelpers';
+import { useRef, useEffect, useState } from 'react';
+import LayerSelector from './LayerSelector.jsx';
+import Map from '@arcgis/core/Map';
+import MapView from '@arcgis/core/views/MapView';
+import FeatureLayer from '@arcgis/core/layers/FeatureLayer';
+import TileInfo from '@arcgis/core/layers/support/TileInfo';
+import WebTileLayer from '@arcgis/core/layers/WebTileLayer';
+import TileLayer from '@arcgis/core/layers/TileLayer';
+import '@arcgis/core/assets/esri/themes/light/main.css';
 
 export default {
   title: 'LayerSelector/WithMap',
-  decorators: [ModulesHelper],
+  component: LayerSelector,
 };
 
-const WithMap = ({ modules, center, zoom, scale, baseLayers, overlays }) => {
+const WithMap = ({ center, zoom, scale, baseLayers, overlays }) => {
   const mapDiv = useRef();
   const [layerSelectorOptions, setLayerSelectorOptions] = useState();
 
   useEffect(() => {
     const initMap = () => {
       console.log('init');
-      const map = new modules.Map();
-      const view = new modules.MapView({
+      const map = new Map();
+      const view = new MapView({
         container: mapDiv.current,
         map,
         center: center ? center : [-112, 40],
@@ -25,18 +31,17 @@ const WithMap = ({ modules, center, zoom, scale, baseLayers, overlays }) => {
 
       setLayerSelectorOptions({
         view: view,
-        quadWord: process.env.QUAD_WORD,
+        quadWord: import.meta.env.VITE_QUAD_WORD,
         baseLayers: baseLayers
           ? baseLayers
           : ['Hybrid', 'Lite', 'Terrain', 'Topo', 'Color IR'],
         overlays: overlays ? overlays : ['Address Points'],
-        modules,
         position: 'top-right',
       });
     };
 
     initMap();
-  }, [modules, zoom, center, scale, baseLayers, overlays]);
+  }, [zoom, center, scale, baseLayers, overlays]);
 
   return (
     <div
@@ -50,16 +55,13 @@ const WithMap = ({ modules, center, zoom, scale, baseLayers, overlays }) => {
   );
 };
 
-export const zoom = (modules) => <WithMap modules={modules} zoom={6} />;
-export const scale = (modules) => <WithMap modules={modules} scale={12000} />;
-export const noZoomOrScale = (modules) => <WithMap modules={modules} />;
-export const zoomAndScale = (modules) => (
-  <WithMap modules={modules} zoom={6} scale={12000} />
-);
-export const customLOD = (modules) => {
+export const zoom = () => <WithMap zoom={6} />;
+export const scale = () => <WithMap scale={12000} />;
+export const noZoomOrScale = () => <WithMap />;
+export const zoomAndScale = () => <WithMap zoom={6} scale={12000} />;
+export const customLOD = () => {
   const tileSize = 256;
   const earthCircumference = 40075016.685568;
-  const halfEarthCircumference = halfEarthCircumference * 0.5;
   const inchesPerMeter = 39.37;
   const initialResolution = earthCircumference / tileSize;
 
@@ -76,13 +78,13 @@ export const customLOD = (modules) => {
 
   const baseLayers = [
     {
-      Factory: modules.WebTileLayer,
+      Factory: WebTileLayer,
       urlTemplate:
         'http://{subDomain}.tile.stamen.com/toner/{level}/{col}/{row}.png',
       id: 'Imagery',
       subDomains: ['a', 'b', 'c', 'd'],
       selected: true,
-      tileInfo: new modules.TileInfo({
+      tileInfo: new TileInfo({
         dpi: 96,
         rows: 256,
         cols: 256,
@@ -98,18 +100,18 @@ export const customLOD = (modules) => {
     },
   ];
 
-  return <WithMap modules={modules} zoom={6} baseLayers={baseLayers} />;
+  return <WithMap zoom={6} baseLayers={baseLayers} />;
 };
-export const linkedOverlays = (modules) => {
+export const linkedOverlays = () => {
   const baseLayers = [
     {
-      Factory: modules.TileLayer,
+      Factory: TileLayer,
       url: 'https://services.arcgisonline.com/ArcGIS/rest/services/World_Topo_Map/MapServer',
       id: 'Topo',
       linked: ['Cities'],
     },
     {
-      Factory: modules.TileLayer,
+      Factory: TileLayer,
       url: 'https://services.arcgisonline.com/ArcGIS/rest/services/World_Terrain_Base/MapServer',
       id: 'Terrain',
       linked: ['Counties'],
@@ -118,33 +120,26 @@ export const linkedOverlays = (modules) => {
 
   const overlays = [
     {
-      Factory: modules.FeatureLayer,
+      Factory: FeatureLayer,
       url: 'https://services1.arcgis.com/99lidPhWCzftIe9K/arcgis/rest/services/UtahMunicipalBoundaries/FeatureServer/0',
       id: 'Cities',
     },
     {
-      Factory: modules.FeatureLayer,
+      Factory: FeatureLayer,
       url: 'https://services1.arcgis.com/99lidPhWCzftIe9K/arcgis/rest/services/UtahCountyBoundaries/FeatureServer/0',
       id: 'Counties',
     },
   ];
 
-  return (
-    <WithMap
-      modules={modules}
-      zoom={6}
-      baseLayers={baseLayers}
-      overlays={overlays}
-    />
-  );
+  return <WithMap zoom={6} baseLayers={baseLayers} overlays={overlays} />;
 };
 
-export const landOwnership = (modules) => {
+export const landOwnership = () => {
   const landOwnershipOptions = {
     overlays: [
       'Address Points',
       {
-        Factory: modules.FeatureLayer,
+        Factory: FeatureLayer,
         url: 'https://gis.trustlands.utah.gov/server/rest/services/Ownership/UT_SITLA_Ownership_LandOwnership_WM/FeatureServer/0',
         id: 'Land Ownership',
         opacity: 0.3,
@@ -153,5 +148,5 @@ export const landOwnership = (modules) => {
     zoom: 10,
   };
 
-  return <WithMap modules={modules} {...landOwnershipOptions} />;
+  return <WithMap {...landOwnershipOptions} />;
 };
