+++ packages/mouse-trap/package.json
@@ -33,10 +33,10 @@
     "/dist"
   ],
   "scripts": {
-    "test": "npm test"
+    "test": "vitest --config ../../vite.config.js"
   },
   "dependencies": {
-    "@arcgis/core": "^4.18.1",
+    "@arcgis/core": "^4.0.0",
     "@ugrc/utilities": "^1.3.1",
     "lodash.throttle": "^4.1.1"
   },
