+++ packages/mouse-trap/src/index.js
@@ -1,3 +1,3 @@
-import useMouseTrap from './MouseTrap';
+import useMouseTrap from './MouseTrap.jsx';
 
 export default useMouseTrap;
