+++ packages/sherlock/package.json
@@ -27,16 +27,17 @@
     "/dist"
   ],
   "scripts": {
-    "test": "npm test"
+    "test": "vitest --config ../../vite.config.js"
   },
   "dependencies": {
+    "@arcgis/core": "^4.0.0",
     "@fortawesome/fontawesome-svg-core": "^1.2.34",
     "@fortawesome/free-solid-svg-icons": "^5.15.2",
     "@fortawesome/react-fontawesome": "^0.1.14",
     "@ugrc/utilities": "^1.3.1",
-    "downshift": "^6.0.15",
-    "lodash.debounce": "^4.0.8",
+    "downshift": "^6.1.7",
     "lodash.escaperegexp": "^4.1.2",
+    "lodash.sortby": "^4.7.0",
     "lodash.uniqwith": "^4.5.0",
     "react-fast-compare": "^3.2.0",
     "reactstrap": "^8.8.1"
