+++ vite.config.js
@@ -11,14 +11,14 @@ module.exports = defineConfig({
       fileName: (format) => `${packageName}.${format}.js`,
     },
     rollupOptions: {
-      // // make sure to externalize deps that shouldn't be bundled into your library
-      // external: ['react'],
-      // output: {
-      //   // Provide global variables to use in the UMD build for externalized deps
-      //   globals: {
-      //     react: 'React',
-      //   },
-      // },
+      // make sure to externalize deps that shouldn't be bundled into your library
+      external: ['react'],
+      output: {
+        // Provide global variables to use in the UMD build for externalized deps
+        globals: {
+          react: 'React',
+        },
+      },
     },
   },
 });
